+++ dll/shellext/netshell/CMakeLists.txt
@@ -11,6 +11,11 @@ endif()
 remove_definitions(-D_WIN32_WINNT=0x502)
 add_definitions(-D_WIN32_WINNT=0x600)
 
+add_definitions(
+    -D_ATL_NO_EXCEPTIONS)
+
+include_directories(${REACTOS_SOURCE_DIR}/sdk/lib/atl)
+
 add_definitions(-D_NETSHELL_)
 spec2def(netshell.dll netshell.spec)
 
@@ -18,7 +23,6 @@ list(APPEND SOURCE
     netshell.cpp
     shfldr_netconnect.cpp
     enumlist.cpp
-    classfactory.cpp
     connectmanager.cpp
     lanconnectui.cpp
     lanstatusui.cpp
@@ -34,8 +38,9 @@ add_library(netshell SHARED
     ${CMAKE_CURRENT_BINARY_DIR}/netshell_stubs.c
     ${CMAKE_CURRENT_BINARY_DIR}/netshell.def)
 
-set_module_type(netshell win32dll)
-target_link_libraries(netshell uuid wine)
-add_importlibs(netshell shlwapi shell32 version iphlpapi ole32 gdi32 user32 advapi32 setupapi ws2_32 comctl32 msvcrt kernel32 ntdll)
+set_module_type(netshell win32dll UNICODE)
+target_link_libraries(netshell atlnew uuid wine)
+add_delay_importlibs(netshell ole32 oleaut32 shlwapi shell32)
+add_importlibs(netshell version iphlpapi gdi32 user32 advapi32 setupapi ws2_32 comctl32 msvcrt kernel32 ntdll)
 add_pch(netshell precomp.h SOURCE)
 add_cd_file(TARGET netshell DESTINATION reactos/system32 FOR all)
